Move _omc_SUNMatScaleIAdd_Sparse to sundials_util
