+++ OMCompiler/SimulationRuntime/c/simulation/solver/ida_solver.h
@@ -37,8 +37,6 @@
 #include "openmodelica.h"
 #include "simulation_data.h"
 #include "util/simulation_options.h"
-#include "sundials_error.h"
-#include "sundials_util.h"
 #include "simulation/solver/solver_main.h"
 #include "omc_config.h" /* for WITH_SUNDIALS */
